+++ ntex-io/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [0.1.0-b.3] - 2021-12-xx
+
+* Rename .poll_write_ready() to .poll_flush()
+
+* Rename .write_ready() to .flush()
+
 ## [0.1.0-b.2] - 2021-12-20
 
 * Removed `WriteRef` and `ReadRef`